math: make ceil return an f64

+++ vlib/math/math.v
@@ -45,8 +45,8 @@ pub fn cbrt(a f64) f64 {
 	return C.cbrt(a)
 }
 
-// ceil returns the nearest integer greater or equal to the provided value.
-pub fn ceil(a f64) int {
+// ceil returns the nearest f64 greater or equal to the provided value.
+pub fn ceil(a f64) f64 {
 	return C.ceil(a)
 }
 
@@ -146,7 +146,7 @@ pub fn factorial(n int) i64 {
 }
 */
 
-// floor returns the nearest integer lower or equal of the provided value.
+// floor returns the nearest f64 lower or equal of the provided value.
 pub fn floor(a f64) f64 {
 	return C.floor(a)
 }
